+++ lnhpdWebApi/lnhpdWebApi.csproj
@@ -284,6 +284,7 @@
     <Compile Include="Models\Response\Pagination.cs" />
     <Compile Include="Models\Response\Response.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Utils\ResponseHelper.cs" />
   </ItemGroup>
   <ItemGroup>
     <Content Include="Global.asax" />
